feat: 예외처리 추가
카테고리 삭제시 게시글 존재하면 예외처리

+++ src/main/java/soonmap/controller/AdminController.java
@@ -276,6 +276,10 @@ public class AdminController {
 
         ArticleType articleType = articleTypeService.findOneById(id);
 
+        if (!articleService.findAllByArticleTypeId(articleType.getId()).isEmpty()) {
+            throw new CustomException(HttpStatus.BAD_REQUEST,"게시글이 존재하는 카테고리입니다.");
+        }
+
         Long deleteById = articleTypeService.deleteById(articleType.getId());
 
         return ResponseEntity.ok()
